requirements.txt
Django==2.2.4
django-cors-headers==3.0.2
djangorestframework==3.10.2
pytz==2019.2
sqlparse==0.3.0
xlrd==1.2.0
